Move fluid_event_from_midi_event() to fluid_event.c
