+++ include/fluidsynth/seqbind.h
@@ -35,8 +35,6 @@ extern "C" {
 FLUIDSYNTH_API
 fluid_seq_id_t fluid_sequencer_register_fluidsynth(fluid_sequencer_t *seq, fluid_synth_t *synth);
 FLUIDSYNTH_API 
-int fluid_event_from_midi_event(fluid_event_t *, const fluid_midi_event_t *);
-FLUIDSYNTH_API 
 int fluid_sequencer_add_midi_event_to_buffer(void *data, fluid_midi_event_t *event);
 /* @} */
