+++ src/midi/fluid_seqbind.c
@@ -357,88 +357,6 @@ static fluid_seq_id_t get_fluidsynth_dest(fluid_sequencer_t *seq)
 
 /**
  * Transforms an incoming MIDI event (from a MIDI driver or MIDI router) to a
- * sequencer event.
- *
- * @param evt returned sequencer event
- * @param event MIDI event
- * @return #FLUID_OK or #FLUID_FAILED
- *
- * @note This function copies the fields of the MIDI event into the provided
- * sequencer event. Calling applications must create the sequencer event and set
- * additional fields such as the source and destination of the sequencer event.
- *
- * @code{.cpp}
- * // ... get MIDI event, e.g. using player_callback()
- *
- * // Send MIDI event to sequencer to play
- * fluid_event_t *evt = new_fluid_event();
- * fluid_event_set_source(evt, -1);
- * fluid_event_set_dest(evt, seqid);
- * fluid_event_from_midi_event(evt, event);
- * fluid_sequencer_send_at(sequencer, evt, 50, 0); // relative time
- * delete_fluid_event(evt);
- * @endcode
- *
- * @since 2.2.7
- */
-int fluid_event_from_midi_event(fluid_event_t *evt, const fluid_midi_event_t *event)
-{
-    fluid_return_val_if_fail(event != NULL, FLUID_FAILED);
-
-    int chan = fluid_midi_event_get_channel(event);
-
-    switch (fluid_midi_event_get_type(event))
-    {
-        case NOTE_OFF:
-            fluid_event_noteoff(evt, chan, (short)fluid_midi_event_get_key(event));
-            break;
-
-        case NOTE_ON:
-            fluid_event_noteon(evt,
-                               fluid_midi_event_get_channel(event),
-                               (short)fluid_midi_event_get_key(event),
-                               (short)fluid_midi_event_get_velocity(event));
-            break;
-
-        case CONTROL_CHANGE:
-            fluid_event_control_change(evt,
-                                       chan,
-                                       (short)fluid_midi_event_get_control(event),
-                                       (short)fluid_midi_event_get_value(event));
-            break;
-
-        case PROGRAM_CHANGE:
-            fluid_event_program_change(evt, chan, (short)fluid_midi_event_get_program(event));
-            break;
-
-        case PITCH_BEND:
-            fluid_event_pitch_bend(evt, chan, fluid_midi_event_get_pitch(event));
-            break;
-
-        case CHANNEL_PRESSURE:
-            fluid_event_channel_pressure(evt, chan, (short)fluid_midi_event_get_program(event));
-            break;
-
-        case KEY_PRESSURE:
-            fluid_event_key_pressure(evt,
-                                     chan,
-                                     (short)fluid_midi_event_get_key(event),
-                                     (short)fluid_midi_event_get_value(event));
-            break;
-
-        case MIDI_SYSTEM_RESET:
-            fluid_event_system_reset(evt);
-            break;
-
-        default: /* Not yet implemented */
-            return FLUID_FAILED;
-    }
-
-    return FLUID_OK;
-}
-
-/**
- * Transforms an incoming MIDI event (from a MIDI driver or MIDI router) to a
  * sequencer event and adds it to the sequencer queue for sending as soon as possible.
  *
  * @param data The sequencer, must be a valid #fluid_sequencer_t
